charms/jupyter-ui/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via httpcore
attrs==22.1.0
    # via jsonschema
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==3.0.4
    # via requests
charmed-kubeflow-chisme==0.0.5
    # via -r requirements.in
deepdiff==6.2.2
    # via charmed-kubeflow-chisme
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.2
    # via lightkube
idna==2.10
    # via
    #   anyio
    #   requests
    #   rfc3986
jinja2==3.1.2
    # via charmed-kubeflow-chisme
jsonschema==3.2
    # via serialized-data-interface
lightkube==0.12.0
    # via charmed-kubeflow-chisme
lightkube-models==1.26.0.4
    # via lightkube
markupsafe==2.1.1
    # via jinja2
oci-image==1.0.0
    # via -r requirements.in
ops==1.5.4
    # via
    #   -r requirements.in
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
pyrsistent==0.19.2
    # via jsonschema
pyyaml==5.4
    # via
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.25
    # via serialized-data-interface
rfc3986[idna2008]==1.5.0
    # via httpx
ruamel-yaml==0.17.21
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
serialized-data-interface==0.3.5
    # via -r requirements.in
six==1.16.0
    # via jsonschema
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
urllib3==1.26.13
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
